Relocate remaining LAN specific security class into lan module
